move courses validation to courses model
