+++ middleware/validator.js
@@ -4,7 +4,9 @@ const validateRole = function(req, res, next) {
     : null;
 
   if (role !== "admin") {
-    return res.status(403).send("Invalid credentials to create a course");
+    return res
+      .status(403)
+      .send("Invalid credentials to create a document on db");
   }
   next();
 };
