+++ models/user.js
@@ -63,31 +63,30 @@ function validateUser(user) {
   return Joi.validate(user, schema);
 }
 
-async function validateUserCourses(courses) {
-  if (!courses) {
-    return [];
-  }
+// async function validateUserCourses(courses) {
+//   if (!courses) {
+//     return [];
+//   }
 
-  let coursesPromises = courses.map(async element => {
-    const course = await Course.findById(element.courseId);
-    if (!course) {
-      let error = {};
-      error.details = [];
-      error.details.push({ message: "Invalid course" });
-      throw error;
-    }
-    const saveCourse = {
-      _id: course._id,
-      name: course.name
-    };
-    return saveCourse;
-  });
+//   let coursesPromises = courses.map(async element => {
+//     const course = await Course.findById(element.courseId);
+//     if (!course) {
+//       let error = {};
+//       error.details = [];
+//       error.details.push({ message: "Invalid course" });
+//       throw error;
+//     }
+//     const saveCourse = {
+//       _id: course._id,
+//       name: course.name
+//     };
+//     return saveCourse;
+//   });
 
-  courses = await Promise.all(coursesPromises);
-  return courses;
-}
+//   courses = await Promise.all(coursesPromises);
+//   return courses;
+// }
 
 exports.User = User;
 exports.validate = validateUser;
 exports.userSchema = userSchema;
-exports.validateUserCourses = validateUserCourses;
